+++ R/backend-oracle.R
@@ -88,7 +88,7 @@ sql_query_upsert.Oracle <- function(con,
 
   insert_cols <- c(by, update_cols)
   insert_cols_esc <- escape(ident(insert_cols), parens = FALSE, con = con)
-  insert_values <- sql_table_prefix(con, insert_cols, ident("...y"))
+  insert_values <- sql_table_prefix(con, insert_cols, "...y")
 
   clauses <- list(
     sql_clause("MERGE INTO", table),
@@ -182,6 +182,10 @@ sql_expr_matches.Oracle <- function(con, x, y, ...) {
   glue_sql2(con, "decode({x}, {y}, 0, 1) = 0")
 }
 
+#' @export
+db_supports_table_alias_with_as.Oracle <- function(con) {
+  FALSE
+}
 
 # roacle package ----------------------------------------------------------
 
@@ -215,4 +219,7 @@ setdiff.OraConnection <- setdiff.tbl_Oracle
 #' @export
 sql_expr_matches.OraConnection <- sql_expr_matches.Oracle
 
-globalVariables(c("DATE", "CURRENT_TIMESTAMP", "TRUNC", "dbms_random.VALUE"))
+#' @export
+db_supports_table_alias_with_as.OraConnection <- db_supports_table_alias_with_as.Oracle
+
+utils::globalVariables(c("DATE", "CURRENT_TIMESTAMP", "TRUNC", "dbms_random.VALUE"))