+++ R/build-sql.R
@@ -97,14 +97,9 @@ sql_quote_transformer <- function(connection) {
     glue_check_collapse(type, should_collapse)
 
     if (type == "tbl") {
-      if (is_bare_character(value)) {
-        value <- ident(value)
-      }
+      value <- as_table_ident(value)
     } else if (type == "from") {
-      # TODO maybe this could call dbplyr_sql_subquery()
-      if (is_bare_character(value)) {
-        value <- ident(value)
-      }
+      value <- as_from(value)
     } else if (type == "col") {
       if (is_bare_character(value)) {
         value <- ident(value)
@@ -137,12 +132,11 @@ sql_quote_transformer <- function(connection) {
 }
 
 glue_check_collapse <- function(type, collapse) {
-  # collapse is only allowed for type `col`
   if (type %in% c("col", "val")) {
     return()
   }
 
   if (collapse) {
-    cli_abort("Collapsing is only allowed for {.val col}, not for {.val {type}}.")
+    cli_abort("Collapsing is only allowed for {.val col} and {.val val}, not for {.val {type}}.")
   }
 }