+++ tests/testthat/test-tbl-lazy.R
@@ -40,3 +40,8 @@ test_that("base source of tbl_lazy is always 'df'", {
   out <- lazy_frame(x = 1, y = 5) %>% sql_build()
   expect_equal(out, base_query(ident("df")))
 })
+
+test_that("names() inform that they aren't meant to be used", {
+  skip("Don't know how to not inform during assignment")
+  expect_snapshot(names(lazy_frame(x = 1)))
+})